requirements.txt
python-telegram-bot==20.3
aiohttp==3.10.10
psycopg2-binary